#32: refactoring
